Move common code to base class
