+++ common/src/codegen/playwright-csharp-mstest/templateCollection.ts
@@ -6,13 +6,21 @@ import { Indent } from "../../file-defs";
 import { EOL } from "os";
 import { indentCharMap } from "../../file-defs/shared";
 
+/** Wrapper for the parameters needed for initializing a TemplateCollection.  */
 export type TemplateCollectionOptions = {
+  /** The default templates directory of the IDE. e.g For MsTest, it's `src\codegen\playwright-csharp-mstest\templates`*/
   templatesDir: string;
+  /** The custom template directory of current RM prroject. This should be `custom-code/templates` */
   customTemplatesDir: string;
+  /** The file extension of the template files. Currently it's `.hbs` */
   fileExtension: string;
+  /** The Indent char that is specified in the .rmproj file */
   requiredIndentChar: Indent;
+  /** The Indent size that is specified in the .rmproj file */
   requiredIndentSize: number;
+  /** The Indent char that is applied in the source .hbs file */
   sourceIndentChar: Indent;
+  /** The Indent size that is applied in the source .hbs file */
   sourceIndentSize: number;
 };
 
@@ -38,31 +46,8 @@ export class MsTestTemplateCollection extends BaseDotnetTemplateCollection imple
   public readonly RUNSETTINGS_FILE: HandlebarsTemplateDelegate<any>;
   public readonly USINGS_FILE: HandlebarsTemplateDelegate<any>;
 
-  private readonly templateLoader: TemplateLoader;
-  private readonly templateCompiler: TemplateCompiler;
-  private readonly templateReindenter: TemplateReindenter;
-
   constructor(templateCollectionOptions: TemplateCollectionOptions) {
-    const {
-      templatesDir,
-      customTemplatesDir,
-      fileExtension,
-      requiredIndentChar,
-      requiredIndentSize,
-      sourceIndentChar,
-      sourceIndentSize,
-    } = templateCollectionOptions;
-
-    super(templatesDir, customTemplatesDir, fileExtension);
-
-    this.templateLoader = new TemplateLoader(templatesDir, customTemplatesDir, fileExtension);
-    this.templateCompiler = new TemplateCompiler();
-    this.templateReindenter = new TemplateReindenter(
-      sourceIndentChar,
-      sourceIndentSize,
-      requiredIndentChar,
-      requiredIndentSize
-    );
+    super(templateCollectionOptions);
 
     this.BASE_CLASSES_FILE = compile("");
     this.COMMENT = this.loadAndCompile("Comment");
@@ -83,76 +68,5 @@ export class MsTestTemplateCollection extends BaseDotnetTemplateCollection imple
     this.CSPROJECT_FILE = this.loadAndCompile("CsProjectFile");
     this.RUNSETTINGS_FILE = this.loadAndCompile("RunSettingsFile");
     this.USINGS_FILE = this.loadAndCompile("UsingsFile");
-  }
-
-  private loadAndCompile(templateFileName: string) {
-    const raw = this.templateLoader.load(templateFileName);
-    const reindented = this.templateReindenter.reindentContent(raw);
-    return this.templateCompiler.compile(reindented);
-  }
-}
-
-export class TemplateReindenter {
-  private readonly sourceIndentChar: Indent;
-  private readonly sourceIndentSize: number;
-  private readonly requiredIndenChar: Indent;
-  private readonly requiredIndentSize: number;
-
-  constructor(
-    sourceIndentChar: Indent,
-    sourceIndentSize: number,
-    requiredIndenChar: Indent,
-    requiredIndentSize: number
-  ) {
-    this.sourceIndentChar = sourceIndentChar;
-    this.sourceIndentSize = sourceIndentSize;
-    this.requiredIndenChar = requiredIndenChar;
-    this.requiredIndentSize = requiredIndentSize;
-  }
-
-  public reindentContent(rawStr: string): string {
-    const lines = rawStr.split(/\r?\n/);
-    const indentedLines = lines.map((l) => this.reindentLine(l));
-    const indented = indentedLines.join(EOL);
-    return indented;
-  }
-
-  private reindentLine(line: string): string {
-    const leadingIndentCharCount = this.countLeadingIndentChars(line);
-    const indentCount = Math.round(leadingIndentCharCount / this.sourceIndentSize);
-    const requiredIndent = indentCharMap.get(this.requiredIndenChar)!.repeat(this.requiredIndentSize * indentCount);
-    const trimmedStr = this.sourceIndentChar === Indent.Spaces ? line.trim() : line.replace(/^\t+/, "");
-    return `${requiredIndent}${trimmedStr}`;
-  }
-
-  private countLeadingIndentChars(str: string): number {
-    const regex = this.sourceIndentChar === Indent.Spaces ? /^\s+/ : /^\t+/;
-    var match = str.match(regex);
-    var count = match ? match[0].length : 0;
-    return count;
-  }
-}
-
-class TemplateLoader {
-  private templatesDir: string;
-  private customTemplatesDir: string;
-  private fileExtension: string;
-
-  constructor(templatesDir: string, customTemplatesDir: string, fileExtension: string) {
-    this.templatesDir = templatesDir;
-    this.customTemplatesDir = customTemplatesDir;
-    this.fileExtension = fileExtension;
-  }
-
-  load(templateFileName: string) {
-    return loadTemplate(this.templatesDir, this.customTemplatesDir, templateFileName + this.fileExtension);
-  }
-}
-
-class TemplateCompiler {
-  constructor() {}
-
-  compile(rawTemplateStr: string): HandlebarsTemplateDelegate<any> {
-    return compile(rawTemplateStr);
   }
 }
